Model Deployment (App)/flask/requirements.txt
pynrrd==1.0.0
pillow==10.2.0
Flask==3.0.3
Flask-Cors==4.0.0
numpy==1.24.3
tensorflow==2.13.1
opencv-python==4.9.0.80
simpleitk==2.3.1
matplotlib==3.8.3
